ImGui support, basic ui features

+++ VoxelEngine/src/Terrain/Chunk.hpp
@@ -39,7 +39,7 @@ public:
 	void Generate();
 	void GenerateMesh();
 	void GenerateEdgeMesh(VoxelFace face);
-	std::pair<Position2D, Position3D> GetPositionRelativeToWorld(glm::vec3 pos) const;
+	std::pair<Position2D, Position3D> GetPositionRelativeToWorld(glm::i16vec3 pos) const;
 
 	VoxelGrid& GetVoxelGrid();
 	const std::vector<Vertex>& GetMesh() const;
